+++ src/main/java/racingcar/ConsoleApplication.java
@@ -1,11 +1,10 @@
 package racingcar;
 
 import racingcar.controller.ConsoleController;
-import racingcar.view.input.InputView;
-import racingcar.view.output.OutputView;
+import racingcar.view.InputView;
+import racingcar.view.OutputView;
 
 public class ConsoleApplication {
-
     public static void main(String[] args) {
         ConsoleController consoleController = new ConsoleController();
         consoleController.execute(new InputView(), new OutputView());
