+++ src/main/java/racingcar/view/InputView.java
@@ -1,4 +1,4 @@
-package racingcar.view.input;
+package racingcar.view;
 
 import java.util.Scanner;
